% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_reshape_meteo.R
\name{complete_meteo}
\alias{complete_meteo}
\title{Complete missing meteo variables}
\usage{
complete_meteo(meteo)
}
\arguments{
\item{meteo}{meteoland meteo data}
}
\value{
the same \code{meteo} data provided with the the variables completed
}
\description{
Calculates missing meteo variables
}
\details{
This function takes a meteo object (with meteoland names) and complete any
missing variable if it is possible
}
\examples{

# example data
data("meteoland_meteo_example")

# remove MinRelativeHumidity
meteoland_meteo_example$MinRelativeHumidity <- NULL
# complete vars
completed_meteo <- complete_meteo(meteoland_meteo_example)
# check MinRelativeHumidity
completed_meteo$MinRelativeHumidity

}
