\encoding{UTF-8}
\name{defaultGenerationParams}
\alias{defaultGenerationParams}
\title{
Default generation parameters
}
\description{
Returns a list with the default parameterization for weather generation.
}
\usage{
defaultGenerationParams()
}
\value{
A list with the following items (default values in brackets):
\itemize{
\item{\code{conditional [= "input"]}: A logical flag to indicate that weather generation is conditional on simulation of annual precipitation.}
\item{\code{dry_wet_threshold [= 0.3]}: Precipitation threshold (mm) for separating dry from wet days.}
\item{\code{wet_extreme_quantile_threshold  [= 0.8]}: Quantile for separating wet from extremely wet days.}
\item{\code{n_knn_annual [= 100]}: Number of years re-sampled using K-nearest neighbour.}
\item{\code{adjust_annual_precip [= TRUE]}: A logical flag to indicate that annual precipitation generated by the algorithm should be adjusted to fit either overall observed precipitation or simulated annual precipitation.}
\item{\code{min_ratio [= 0.9]}: Minimum adjustment ratio for precipitation.}
\item{\code{max_ratio [= 1.2]}: Minimum adjustment ratio for precipitation.}
}
}
\references{
Steinschneider S. & Brown C. (2013) A semiparametric multivariate, multisite weather generator with low-frequency variability for use in climate risk assessments. Water Resources Research 49, 7205-7220. doi:10.1002/wrcr.20528.
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CTFC
}
\seealso{\code{\link{weathergeneration}}}