% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weathergeneration.R
\name{.mc_assign_states}
\alias{.mc_assign_states}
\title{Assign Markov States from Precipitation and State Thresholds}
\usage{
.mc_assign_states(x, months, thresholds)
}
\arguments{
\item{x}{vector of daily precipitation}

\item{months}{vector of months corresponding to daily precipitation}

\item{thresholds}{list of monthly transition matrices generated from
mc_fit()}

\item{states}{character list of markov states}
}
\description{
Assign Markov States from Precipitation and State Thresholds
}
