% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weathergeneration.R
\name{.mc_sim}
\alias{.mc_sim}
\title{Run Markov Chain Simulation}
\usage{
.mc_sim(months, initial, transitions)
}
\arguments{
\item{months}{numeric array of months for each daily time step}

\item{initial}{initial state}

\item{transitions}{monthly list of transition matrices generated by mc_fit()}
}
\description{
Run Markov Chain Simulation
}
\examples{
transitions <- mc_fit(x=sample(c('d', 'w', 'e'), size=720, replace=TRUE, prob=c(0.5, 0.3, 0.2)), 
                      months=rep(rep(seq(1, 12), each=30), times=2))
mc_simulate(months=rep(1:12, each=30), initial='d', transitions=transitions, states=c('d', 'w', 'e'))

}
