% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weathergeneration.R
\name{.mc_state_threshold}
\alias{.mc_state_threshold}
\title{Determine Markov State Thresholds from Precipitation}
\usage{
.mc_state_threshold(
  x,
  months,
  dry_wet_threshold = 0.3,
  wet_extreme_quantile_threshold = 0.8
)
}
\arguments{
\item{x}{vector of daily precipitation}

\item{months}{vector of months corresponding to daily precipitation vector \code{x}}

\item{dry_wet_threshold}{threshold precipitation amount for dry/wet states}

\item{wet_extreme_quantile_threshold}{threshold quantile for wet/extreme states}
}
\value{
monthly list of precipitation thresholds as vectors of length 2
(first element is threshold amount between dry/wet,
second element is threshold amount between wet/extreme)
}
\description{
Determine Markov State Thresholds from Precipitation
}
