\encoding{UTF-8}
\name{exampledownscalingdata}
\alias{exampledownscalingdata}
\docType{data}
\title{
Example data set for statistical downscaling of RCM predictions
}
\description{
  Example data set including the predictions of Regional Climate Model (CCLM4-8-17; driving global model CNRM-CERFACS-CNRM-CM5) for 12 model cells in a small area in Catalonia (NE Spain). Meteorological data covers an historical period (2000-2003) and a future period (2020-2023), the latter simulated under rcp4.5 scenario.
}
\usage{data("exampledownscalingdata")}
\format{
   Formal class '\code{\link{MeteorologyDownscalingData-class}}'
}
\source{
 ESFG web site (http://esgf.llnl.gov/) that centralizes climate data from GCM and RCM uploaded in the frame of different international consortium, including the EURO-CORDEX regionalisation project.
}
\examples{
data(exampledownscalingdata)
}
\keyword{datasets}
