% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomatch.R
\docType{data}
\name{examplegridtopography}
\alias{examplegridtopography}
\title{Example spatial grid topography}
\format{
Formal class 'SpatialGridTopography'
}
\source{
'Institut Cartogràfic de Catalunya' (ICC)
}
\description{
'SpatialGridTopography' object describing topographic features for a grid of
5 km x 5 km and cell size of 100 m in Catalonia (NE Spain).
}
\examples{

data(examplegridtopography)

}
\keyword{datasets}
