% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomatch.R
\docType{data}
\name{exampleinterpolationdata}
\alias{exampleinterpolationdata}
\title{Example data set for interpolation from weather stations}
\format{
Formal class '\code{\link{MeteorologyInterpolationData-class}}'
}
\source{
'Servei Meteorològic de Catalunya' (SMC) and 'Agencia Española de
Meteorología' (AEMET)
}
\description{
Example data set of spatial location, topography and daily meteorological
records from 38 weather stations in Catalonia (NE Spain) corresponding to
years 2000-2003.
}
\examples{

data(exampleinterpolationdata)

}
\keyword{datasets}
