% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractdates.R
\name{extractdates}
\alias{extractdates}
\alias{extractvars}
\alias{extractgridindex}
\alias{extractgridpoints}
\title{Extracts meteorological data}
\usage{
extractdates(object, dates = NULL, verbose = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{SpatialPointsMeteorology}},
\code{\link{SpatialGridMeteorology}} or
\code{\link{SpatialPixelsMeteorology}}.}

\item{dates}{A vector of \code{\link{Date}} with a (subset) of dates to be
extracted. If \code{NULL} all dates will be returned.}

\item{verbose}{Boolean flag to print process information.}

\item{vars}{A character vector with the set of variables to be extracted.}

\item{grid}{An object of class \code{\link{SpatialGridMeteorology-class}} or
\code{\link{SpatialPixelsMeteorology-class}} with the meteorological data
for a full grid or a subset of grid cells, respectively. Alternatively, a
string specifying a NetCDF to be read from the disk.}

\item{index}{An integer with a grid index.}

\item{points}{An object of class \code{\link{SpatialPoints}}.}
}
\value{
\itemize{ \itemFunction \code{extractdates()}, returns a \code{list}
with the same length as \code{dates}. Each element of the list is a spatial
object (\code{\link{SpatialPointsDataFrame}},
\code{\link{SpatialGridDataFrame}} or \code{\link{SpatialPixelsDataFrame}},
depending on the input) with the meteorological data for all the spatial
elements. If only one date is asked, the function returns directly the
spatial object, without embedding it into a list. \itemFunction
\code{extractvars()}, returns a \code{list} with the same length as
\code{vars}. Each element of the list is a spatial object
(\code{\link{SpatialPointsDataFrame}}, \code{\link{SpatialGridDataFrame}} or
\code{\link{SpatialPixelsDataFrame}}, depending on the input) with the
meteorological data for all the spatial elements. If only one variable is
asked, the function returns directly the spatial object, without embedding
it into a list. \itemFunction \code{extractgridindex()} returns a data
frame. \itemFunction \code{extractgridpoints()} returns an object of class
\code{\link{SpatialPointsMeteorology}}. }
}
\description{
Extracts meteorological data from an object.
}
\details{
Function \code{extractpoints} is deprecated, because its functionality can
be achieved using subsetting of spatial classes
\code{\link{SpatialGridMeteorology}} and
\code{\link{SpatialPixelsMeteorology}}.
}
\author{
Miquel De \enc{CáceresCaceres} Ainsa, CREAF
}
