% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteocomplete.R
\name{meteocomplete}
\alias{meteocomplete}
\alias{precipitation_concentration}
\alias{precipitation_rainfallErosivity}
\title{Complete daily meteorological variables}
\usage{
meteocomplete(x, latitude, elevation, slope, aspect)
}
\arguments{
\item{x}{A data frame of meteorological variables (dates as row names and
variables as columns).}

\item{latitude}{Latitude in degrees North.}

\item{elevation}{Elevation in m.a.s.l.}

\item{slope}{Slope in degrees.}

\item{aspect}{Aspect in degrees from North.}

\item{p}{A numeric vector with daily precipitation values.}

\item{long}{Longitude in degrees.}

\item{scale}{Either 'month' or 'year'.}

\item{average}{Boolean flag to calculate multi-year averages before applying
MedREM's formula.}
}
\value{
A data frame copied from \code{x} but with filled values for
variables: \itemize{ \item\code{MeanRelativeHumidity}: Mean daily relative
humidity (in percent).  \item\code{MinRelativeHumidity}: Minimum daily
relative humidity (in percent).  \item\code{MaxRelativeHumidity}: Maximum
daily relative humidity (in percent).  \item\code{Radiation}: Incoming solar
radiation (MJ/m2).  \item\code{PET}: Potential evapotranspiration (in mm of
water).  }

Function \code{precipitation_concentration()} returns a value
between 0 (equal distribution of rainfall) and 1 (one day concentrates all
rainfall). Function \code{precipitation_rainfallErosivity()} returns a
vector of twelve values (one for each month) (in MJ·mm·ha-1·h-1·month-1) or
one value (in in MJ·mm·ha-1·h-1·yr-1) depending on the scale.
}
\description{
Fills missing values of relative humidity, radiation and potential
evapotranspiration from a data frame with daily values of
minimum/maximum/mean temperature and precipitation.
}
\details{
The function fills values for humidity, radiation and PET only if they are
missing in the input data frame. If a column 'SpecificHumidity' is present
in the input data, relative humidity is calculated from it. Otherwise,
relative humidity is calculated assuming that dew point temperature equals
the minimum temperature. Potential solar radiation is calculated from
latitude, slope and aspect. Incoming solar radiation is then corrected
following Thornton & Running (1999) and potential evapotranspiration
following Penman (1948).
}
\references{
Thornton, P.E., Running, S.W., 1999. An improved algorithm for
estimating incident daily solar radiation from measurements of temperature,
humidity, and precipitation. Agric. For. Meteorol. 93, 211-228.

Penman, H. L. 1948. Natural evaporation from open water, bare soil and
grass. Proceedings of the Royal Society of London. Series A. Mathematical
and Physical Sciences, 193, 120-145.
Precipitation utils

Function \code{precipitation_concentration()} calculates daily precipitation
concentration (Martin-Vide et al. 2004). Function
\code{precipitation_rainfallErosivity()} calculates a multi-year average of
monthly rainfall erosivity using the MedREM model proposed by Diodato and
Bellochi (2010) for the Mediterranean area (see also Guerra et al. 2016).

MedREM model is: Rm = b0·P·sqrt(d)·(alpha + b1*longitude), where P is
cumulated precipitation and d is maximum daily precipitation. Parameters
used for the MedREM model are b0 = 0.117, b1 = -0.015, alpha = 2. Note that
there is a mistake in Guerra et al. (2016) regarding parameters b1 and a.

Diodato, N., Bellocchi, G., 2010. MedREM, a rainfall erosivity
model for the Mediter-ranean region. J. Hydrol. 387, 119–127,
doi:10.1016/j.jhydrol.2010.04.003.

Guerra CA, Maes J, Geijzendorffer I, Metzger MJ (2016) An assessment of soil
erosion prevention by vegetation in Mediterranean Europe: Current trends of
ecosystem service provision. Ecol Indic 60:213–222. doi:
10.1016/j.ecolind.2015.06.043.

Martin-Vide J (2004) Spatial distribution of a daily precipitation
concentration index in peninsular Spain. International Journal of
Climatology 24, 959–971. doi:10.1002/joc.1030.
}
\seealso{
\code{\link{penman}}, \code{\link{radiation_solarRadiation}}
}
\author{
Miquel De \enc{CáceresCaceres} Ainsa, CREAF

Miquel De \enc{CáceresCaceres} Ainsa, CREAF.
}
