% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteoland-package.R
\docType{package}
\name{meteoland-package}
\alias{meteoland-package}
\alias{meteoland}
\title{meteoland: Landscape Meteorology Tools}
\description{
Functions to estimate weather variables at any position of a landscape
}
\references{
De Caceres et al., 2018 (\doi{10.1016/j.envsoft.2018.08.003}))
}
\seealso{
Useful links: \itemize{ \item{
\url{https://emf-creaf.github.io/meteoland/index.html}} }
}
\author{
\strong{Maintainer}: Miquel De Cáceres
\email{miquelcaceres@gmail.com}
\href{https://orcid.org/0000-0001-7132-2080}{ORCID}

Authors: \itemize{
\item{ Víctor Granda
\href{https://orcid.org/0000-0002-0469-1991}{ORCID}}
\item{ Nicolas Martin
\href{https://orcid.org/0000-0001-7574-0108}{ORCID}}
\item{ Antoine Cabon
\href{https://orcid.org/0000-0001-6426-1726}{ORCID}}
}
}
\keyword{internal}
