% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meteoland_interpolator_example}
\alias{meteoland_interpolator_example}
\title{Example interpolator object}
\format{
stars data cube object
}
\source{
Spanish National Forest Inventory
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Example interpolator with daily meteorological
records from 189 weather stations in Catalonia (NE Spain) corresponding to
April 2022.
}
\examples{

data(meteoland_interpolator_example)

}
\keyword{datasets}
