% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meteoland_meteo_no_topo_example}
\alias{meteoland_meteo_no_topo_example}
\title{Example data set for meteo data from weather stations, without topography}
\format{
sf object
}
\source{
'Servei Meteorològic de Catalunya' (SMC)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Example data set of spatial location and daily meteorological
records from 189 weather stations in Catalonia (NE Spain) corresponding to
April 2022.
}
\examples{

data(meteoland_meteo_no_topo_example)

}
\keyword{datasets}
