% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meteoland_topo_example}
\alias{meteoland_topo_example}
\title{Example data set for topography data from weather stations, without meteo}
\format{
sf object
}
\source{
'Servei Meteorològic de Catalunya' (SMC)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Example data set of spatial location and topography
records from 189 weather stations in Catalonia (NE Spain).
}
\examples{

data(meteoland_topo_example)

}
\keyword{datasets}
