% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineProcedureClasses.R
\docType{class}
\name{MeteorologyInterpolationData-class}
\alias{MeteorologyInterpolationData-class}
\title{Class \code{"MeteorologyInterpolationData"}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

An S4 class to interpolate meteorology over a landscape.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("MeteorologyInterpolationData", ...)}, or by calls to the function
\code{\link{MeteorologyInterpolationData}}.
}

\examples{

#Structure of the S4 object
showClass("MeteorologyInterpolationData")

}
\seealso{
\code{\link{MeteorologyInterpolationData}},
\code{\link{MeteorologyProcedureData-class}}, \code{\link{subsample}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
\keyword{classes}
