% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{points_to_interpolate_example}
\alias{points_to_interpolate_example}
\title{Example data set of points for interpolation of weather variables}
\format{
sf object
}
\source{
Spanish National Forest Inventory
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Example data set of spatial location and topography
records from 15 experimental plots in Catalonia (NE Spain).
}
\examples{

data(points_to_interpolate_example)

}
\keyword{datasets}
