% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitationconcentration.R
\name{precipitation_concentration}
\alias{precipitation_concentration}
\title{Precipitation daily concentration}
\usage{
precipitation_concentration(p)
}
\arguments{
\item{p}{A numeric vector with daily precipitation values.}
}
\value{
Function \code{precipitation_concentration()} returns a value
between 0 (equal distribution of rainfall) and 1 (one day concentrates all
rainfall).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Function \code{precipitation_concentration()} calculates daily precipitation
concentration (Martin-Vide et al. 2004).
}
\references{
Martin-Vide J (2004) Spatial distribution of a daily precipitation
concentration index in peninsular Spain. International Journal of
Climatology 24, 959–971. doi:10.1002/joc.1030.
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF.
}
