\encoding{UTF-8}
\name{precipitationconcentration}
\alias{precipitationconcentration}
\title{
   Daily precipitation concentration
}
\description{
Calculates daily precipitation concentration (Martin-Vide et al. 2004)
}
\usage{
precipitationconcentration(p)
}
\arguments{
  \item{p}{A numeric vector with daily precipitation values.}
}
\value{
An value between 0 (equal distribution of rainfall) and 1 (one day concentrates all rainfall).
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, Centre Tecnologic Forestal de Catalunya
}
\references{
Martin-Vide J (2004) Spatial distribution of a daily precipitation concentration index in peninsular Spain. International Journal of Climatology 24, 959–971. doi:10.1002/joc.1030.
}

