% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raster_to_interpolate_example}
\alias{raster_to_interpolate_example}
\title{Example raster data set for interpolation of weather variables}
\format{
stars object
}
\source{
ICGC
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Example raster data set of spatial location and topography
records from Catalonia (NE Spain). Cell size is 1km x 1km and raster size is 10x10 cells.
}
\examples{

data(raster_to_interpolate_example)

}
\keyword{datasets}
