% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCDFutils.R
\name{readNetCDFpoints}
\alias{readNetCDFpoints}
\alias{readNetCDFdates}
\alias{readNetCDFgridtopology}
\alias{readNetCDFproj4string}
\title{Utility functions for NetCDFs}
\usage{
readNetCDFpoints(file)

readNetCDFdates(file)

readNetCDFgridtopology(file)

readNetCDFproj4string(file)
}
\arguments{
\item{file}{String of the NetCDF whose spatial/temporal coordinates are
desired.}
}
\value{
\itemize{ \item{Function \code{readNetCDFdates} returns a
\code{\link{Date}} vector.} \item{Function \code{readNetCDFpoints} returns an
object \code{\link{SpatialPoints-class}}.} \item{Function
\code{readNetCDFgridtopology} returns an object
\code{\link{GridTopology-class}}.} \item{Function \code{readNetCDFproj4string}
returns an object \code{\link{CRS-class}}.} }
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Functions to read spatial and temporal coordinates from NetCDFs.
}
\section{Functions}{
\itemize{
\item \code{readNetCDFdates()}: \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\item \code{readNetCDFgridtopology()}: \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\item \code{readNetCDFproj4string()}: \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

}}
\seealso{
\code{\link{readmeteorologypoints}},
\code{\link{readmeteorologygrid}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
