% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomatch.R
\docType{class}
\name{SpatialGridMeteorology-class}
\alias{SpatialGridMeteorology-class}
\alias{[,SpatialGridMeteorology,ANY,ANY-method}
\alias{[,SpatialGridMeteorology,ANY,ANY,ANY-method}
\alias{show,SpatialGridMeteorology-method}
\title{Class \code{"SpatialGridMeteorology"}}
\description{
An S4 class that represents a spatial grid with meteorology daily data.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("SpatialGridMeteorology", ...)}, or by calls to the function
\code{\link{SpatialGridMeteorology}}.
}

\examples{

#Structure of the S4 object
showClass("SpatialGridMeteorology")

}
\seealso{
\code{\link{SpatialGridTopography}},
\code{\link{SpatialGridDataFrame-class}}
}
\author{
Miquel De \enc{CáceresCaceres} Ainsa, CREAF
}
\keyword{classes}
