% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomatch.R
\docType{class}
\name{SpatialPixelsTopography-class}
\alias{SpatialPixelsTopography-class}
\alias{[,SpatialPixelsTopography,ANY,ANY-method}
\alias{[,SpatialPixelsTopography,ANY,ANY,ANY-method}
\alias{show,SpatialPixelsTopography-method}
\title{Class \code{"SpatialPixelsTopography"}}
\description{
An S4 class that represents topography that has locations on a regular grid.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("SpatialPixelsTopography", ...)}, or by calls to the function
\code{\link{SpatialPixelsTopography}}.
}

\examples{

#Structure of the S4 object
showClass("SpatialPixelsTopography")

}
\seealso{
\code{\link{SpatialPixelsTopography}},
\code{\link{SpatialPixelsDataFrame-class}}
}
\author{
Miquel De \enc{CáceresCaceres} Ainsa, CREAF
}
\keyword{classes}
