% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineSpatialDataClasses.R
\docType{class}
\name{SpatialPointsMeteorology-class}
\alias{SpatialPointsMeteorology-class}
\title{Class \code{"SpatialPointsMeteorology"}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

An S4 class that represents a set of points with meteorology data series.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("SpatialPointsMeteorology", ...)}, or by calls to the function
\code{\link{SpatialPointsMeteorology}}.
}

\examples{

#Structure of the S4 object
showClass("SpatialPointsMeteorology")

}
\seealso{
\code{\link{SpatialPointsTopography-class}},
\code{SpatialPoints}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
\keyword{classes}
