\encoding{UTF-8}
\name{SpatialPointsMeteorology}
\alias{SpatialPointsMeteorology}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Creates a 'SpatialPointsMeteorology'
}
\description{
Initializes an object of class \code{SpatialPointsMeteorology-class}
}
\usage{
SpatialPointsMeteorology(points, data, dates)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{points}{An object of class \code{\link{SpatialPoints-class}}}
  \item{data}{A vector of data frames (one per spatial point).}
  \item{dates}{Object of class \code{"Date"} describing the time period of meteorological estimates.}
}
\value{An object of class \code{\link{SpatialPointsMeteorology-class}}}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, Centre Tecnologic Forestal de Catalunya
}

\seealso{
\code{\link{SpatialPointsMeteorology-class}}
}
