% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineSpatialDataClasses.R
\docType{class}
\name{SpatialPointsTopography-class}
\alias{SpatialPointsTopography-class}
\title{Class \code{"SpatialPointsTopography"}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

An S4 class that represents topography over a grid of coordinates.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("SpatialPointsTopography", ...)}, or by calls to the function
\code{\link{SpatialPointsTopography}}.
}

\examples{

#Structure of the S4 object
showClass("SpatialPointsTopography")

}
\seealso{
\code{\link{SpatialPointsTopography}},
\code{SpatialPointsDataFrame}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
\keyword{classes}
