% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_meteo.R
\name{write_interpolator}
\alias{write_interpolator}
\title{Write the interpolator object}
\usage{
write_interpolator(interpolator, filename, .overwrite = FALSE)
}
\arguments{
\item{interpolator}{meteoland interpolator object, as created by
\code{\link{create_meteoland_interpolator}}}

\item{filename}{file name for the interpolator nc file}

\item{.overwrite}{logical indicating if the file should be overwrited if it
already exists}
}
\value{
invisible interpolator object, to allow using this function as a step
in a pipe
}
\description{
Write the interpolator object to a file
}
\details{
This function writes the interpolator object created with
\code{\link{create_meteoland_interpolator}} in a NetCDF-CF standard compliant
format, as specified in
http://cfconventions.org/cf-conventions/cf-conventions.html
}
\seealso{
Other interpolator functions: 
\code{\link{add_topo}()},
\code{\link{create_meteo_interpolator}()},
\code{\link{get_interpolation_params}()},
\code{\link{read_interpolator}()},
\code{\link{set_interpolation_params}()},
\code{\link{with_meteo}()}
}
\concept{interpolator functions}
