% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averagearea.R
\name{averagearea}
\alias{averagearea}
\title{Averages area weather}
\usage{
averagearea(object, na.rm = TRUE)
}
\arguments{
\item{object}{An object of class
\code{\link{SpatialPointsMeteorology-class}},
\code{\link{SpatialGridMeteorology-class}} or
\code{\link{SpatialPixelsMeteorology-class}}.}

\item{na.rm}{Boolean flag to indicate that missing values should be excluded
from averages.}
}
\value{
An object of class as the input
\code{SpatialPointsMeteorology-class} with weather series corresponding to
the spatial average of the input.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Averages the weather data series of points or grid pixels.
}
\details{
Assumes that all points/pixels represent the same area.
}
\examples{

\donttest{
data(examplegridtopography)
data(exampleinterpolationdata)

#Interpolation of meteorology over a grid for two days
ml = interpolationgrid(exampleinterpolationdata, examplegridtopography,
                       as.Date(c("2001-02-03", "2001-06-03")))

#Call averaging function
pa = averagearea(ml)

#Spatial information
pa

#Weather data
pa@data[[1]]
}

}
\seealso{
\code{\link{weathergeneration}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
