% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_meteo.R
\name{create_meteo_interpolator}
\alias{create_meteo_interpolator}
\title{Meteoland interpolator creation}
\usage{
create_meteo_interpolator(meteo_with_topo, params = NULL, ...)
}
\arguments{
\item{meteo_with_topo}{meteo object, as returned by \code{\link{with_meteo}}}
}
\value{
an interpolator object (stars)
}
\description{
Function to create the meteoland interpolator
}
\details{
This function takes meteorology information and a list of interpolation
parameters and creates the interpolator object to be ready to use.
}
\seealso{
Other interpolator functions: 
\code{\link{add_topo}()},
\code{\link{read_interpolator}()},
\code{\link{with_meteo}()},
\code{\link{write_interpolator}()}
}
\concept{interpolator functions}
