\encoding{UTF-8}
\name{examplecorrectiondata}
\alias{examplecorrectiondata}
\docType{data}
\title{
Example data set for statistical correction of RCM predictions
}
\description{
  Example data set including the predictions of Regional Climate Model (CCLM4-8-17; driving global model CNRM-CERFACS-CNRM-CM5) for 3 model cells in a small area in Catalonia (NE Spain). Meteorological data covers an historical (reference) period (2000-2003) and a future (projection) period (2020-2023), the latter simulated under rcp4.5 scenario.
}
\usage{data("examplecorrectiondata")}
\format{
   Formal class '\code{\link{MeteorologyUncorrectedData-class}}'
}
\source{
 ESFG web site (http://esgf.llnl.gov/) that centralizes climate data from GCM and RCM uploaded in the frame of different international consortium, including the EURO-CORDEX regionalisation project.
}
\examples{
data(examplecorrectiondata)
}
\keyword{datasets}
