% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplegridtopography}
\alias{examplegridtopography}
\title{Example spatial grid topography}
\format{
Formal class 'SpatialGridTopography'
}
\source{
'Institut Cartogràfic de Catalunya' (ICC)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

'SpatialGridTopography' object describing topographic features for a grid of
5 km x 5 km and cell size of 100 m in Catalonia (NE Spain).
}
\examples{

data(examplegridtopography)

}
\keyword{datasets}
