\encoding{UTF-8}
\name{exampleinterpolationdata}
\alias{exampleinterpolationdata}
\docType{data}
\title{
Example data set for interpolation from weather stations
}
\description{
  Example data set of spatial location, topography and daily meteorological records from 38 weather stations in Catalonia (NE Spain) corresponding to years 2000-2003.
}
\usage{data("exampleinterpolationdata")}
\format{
   Formal class '\code{\link{MeteorologyInterpolationData-class}}'
}
\source{
'Servei Meteorològic de Catalunya' (SMC) and 'Agencia Española de Meteorología' (AEMET)
}
\examples{
data(exampleinterpolationdata)
}
\keyword{datasets}
