% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergegrids.R
\name{mergegrids}
\alias{mergegrids}
\alias{mergepoints}
\title{Merges meteorological data}
\usage{
mergegrids(..., verbose = TRUE)

mergepoints(..., verbose = TRUE)
}
\arguments{
\item{...}{Objects to be merged, either of class
\code{SpatialPointsMeteorology}, \code{SpatialGridMeteorology} or
\code{SpatialPixelsMeteorology}. All objects of the same class.}

\item{verbose}{A logical flag to indicate console output.}
}
\value{
Function codemergepoints returns an object
\code{\link{SpatialPointsMeteorology-class}}. Function \code{mergegrids}
returns an object \code{\link{SpatialGridMeteorology-class}} or an object
\code{\link{SpatialPixelsMeteorology-class}}, depending on the input.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Merges point or gridded meteorological data into a single object
}
\details{
Function \code{mergepoints} requires all coordinate reference systems to the
same. The function allows merging data from the same points (i.e. if they
have the same coordinates). Function \code{mergegrids} pools dates and
variables, but spatial structures (i.e. grid topology, pixel indices,
reference system, ...) should be the same for all objects to be merged.
}
\section{Functions}{
\itemize{
\item \code{mergepoints()}: \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

}}
\seealso{
\code{\link{SpatialPointsMeteorology-class}},
\code{\link{SpatialGridMeteorology-class}},
\code{\link{SpatialPixelsMeteorology-class}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
