% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineProcedureClasses.R
\docType{class}
\name{MeteorologyProcedureData-class}
\alias{MeteorologyProcedureData-class}
\title{Class \code{"MeteorologyProcedureData"}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

A virtual class for estimating meteorology over landscapes
}
\section{Objects from the Class}{
 A virtual Class: No objects may be created
from it.
}

\examples{

showClass("MeteorologyProcedureData")

}
\seealso{
\code{\link{MeteorologyInterpolationData-class}},
\code{\link{MeteorologyUncorrectedData-class}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
\keyword{classes}
