% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineProcedureClasses.R
\docType{class}
\name{MeteorologyUncorrectedData-class}
\alias{MeteorologyUncorrectedData-class}
\title{Class \code{"MeteorologyUncorrectedData"}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

An S4 class to conduct statistical correction of meteorology over a
landscape.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("MeteorologyUncorrectedData", ...)}, or by calls to the function
\code{\link{MeteorologyUncorrectedData}}.
}

\examples{

#Structure of the S4 object
showClass("MeteorologyUncorrectedData")

}
\seealso{
\code{\link{MeteorologyUncorrectedData}},
\code{\link{MeteorologyProcedureData-class}},
\code{\link{examplecorrectiondata}}, \code{\link{subsample}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
\keyword{classes}
