% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialGridMeteorology-methods.R
\name{SpatialGridMeteorology}
\alias{SpatialGridMeteorology}
\title{Creates a 'SpatialGridMeteorology'}
\usage{
SpatialGridMeteorology(grid, proj4string = CRS(as.character(NA)), data, dates)
}
\arguments{
\item{grid}{An object of class \code{\link{GridTopology-class}}}

\item{proj4string}{Object of class \code{"CRS"} with the projection string.}

\item{data}{A vector of data frames (one per date).}

\item{dates}{Object of class \code{"Date"} describing the time period of
meteorological estimates.}
}
\value{
An object of class \code{\link{SpatialGridMeteorology-class}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Initializes an object of class \code{SpatialGridMeteorology-class}
}
\seealso{
\code{\link{SpatialGridMeteorology-class}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
