% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineSpatialDataClasses.R
\docType{class}
\name{SpatialPixelsMeteorology-class}
\alias{SpatialPixelsMeteorology-class}
\title{Class \code{"SpatialPixelsMeteorology"}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

An S4 class that represents meteorology data that has locations on a regular
grid.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("SpatialPixelsMeteorology", ...)}, or by calls to the function
\code{\link{SpatialPixelsMeteorology}}.
}

\examples{

#Structure of the S4 object
showClass("SpatialPixelsMeteorology")

}
\seealso{
\code{\link{SpatialPixelsTopography}},
\code{SpatialPixelsDataFrame}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
\keyword{classes}
