% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPixelsMeteorology-methods.R
\name{SpatialPixelsMeteorology}
\alias{SpatialPixelsMeteorology}
\title{Creates a 'SpatialPixelsMeteorology'}
\usage{
SpatialPixelsMeteorology(
  points,
  data,
  dates,
  tolerance = sqrt(.Machine$double.eps),
  proj4string = as.character(NA),
  round = NULL,
  grid = NULL
)
}
\arguments{
\item{points}{An object of class \code{SpatialPoints}.}

\item{data}{A vector of data frames (one per date).}

\item{dates}{Object of class \code{"Date"} describing the time period of
meteorological estimates.}

\item{tolerance}{Precision up to which extent points should be exactly on a
grid.}

\item{proj4string}{Object of class \code{CRS} in the first
form only used when points does not inherit from
\code{Spatial}.}

\item{round}{default \code{NULL}, otherwise a value passed to as the digits
argument to \code{\link{round}} for setting cell size.}

\item{grid}{Grid topology using an object of class
\code{GridTopology}; a value of \code{NULL} implies that this
will be derived from the point coordinates.}
}
\value{
An object of class \code{\link{SpatialPixelsMeteorology-class}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Initializes an object of class \code{SpatialPixelsMeteorology-class}
}
\seealso{
\code{\link{SpatialPixelsMeteorology-class}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
